refactor clang module infrastructure
